+++ tests/_data/models/modules.php
@@ -1,0 +1,23 @@
+<?php
+
+return [
+    'news' => [
+        'module' => 'news',
+        'name' => 'wdmg/yii2-news',
+        'description' => 'News manager',
+        'class' => 'wdmg\\news\\Module',
+        'bootstrap' => NULL,
+        'homepage' => 'https://github.com/wdmg/yii2-news',
+        'support' => NULL,
+        'authors' => NULL,
+        'require' => NULL,
+        'type' => 'yii2-extension',
+        'license' => NULL,
+        'version' => '1.1.1',
+        'options' => 'a:5:{s:11:"routePrefix";s:5:"admin";s:9:"baseRoute";s:5:"/news";s:10:"baseLayout";s:23:"@app/views/layouts/main";s:9:"imagePath";s:13:"/uploads/news";s:14:"supportLocales";a:3:{i:0;s:5:"ru-RU";i:1;s:5:"uk-UA";i:2;s:5:"en-US";}}',
+        'status' => 1,
+        'protected' => 0,
+        'priority' => 4
+    ]
+];
+