+++ tests/fixtures/ModulesFixture.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace wdmg\news\tests\fixtures;
+
+class ModulesFixture extends \yii\test\ActiveFixture
+{
+    public $modelClass = 'wdmg\admin\models\Modules';
+}